more refactoring, changes that should have been part of the last commint
