+++ pycomm3/socket_.py
@@ -1,6 +1,33 @@
+# -*- coding: utf-8 -*-
+#
+# const.py - A set of structures and constants used to implement the Ethernet/IP protocol
+#
+# Copyright (c) 2019 Ian Ottoway <ian@ottoway.dev>
+# Copyright (c) 2014 Agostino Ruscito <ruscito@gmail.com>
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+#
+
 import socket
 from autologging import logged
-from . import CommError, const
+from . import CommError
+from .const import HEADER_SIZE
 import struct
 
 
@@ -38,7 +65,7 @@ class Socket:
                 self.sock.settimeout(timeout)
             data = self.sock.recv(4096)
             data_len = struct.unpack_from('<H', data, 2)[0]
-            while len(data) - const.HEADER_SIZE < data_len:
+            while len(data) - HEADER_SIZE < data_len:
                 data += self.sock.recv(4096)
 
             return data
